Renamed sidebar_layout.vue file
Renamed to conform with Pascal Case naming of other files
