+++ src/views/layouts/SidebarLayout.vue
